Number of occurrence X in A
